fix multi-assignment projection semantics and build regression
Agent-Logs-Url: https://github.com/gredice/gredice/sessions/2ea0abbb-d499-4fa7-aa53-6a1f44906b32
